장바구니 컴포넌트 분리/ 장바구니 내부 fetch(get)
