+++ src/Components/NavForm/CartForm/CartItem.scss
@@ -1,0 +1,86 @@
+@import '../../../Styles/var.scss';
+
+ul {
+  // @include center();
+  li {
+    display: flex;
+    flex-direction: row;
+    position: relative;
+    width: 100%;
+    height: 78px;
+    padding: 10px 0 30px 0;
+
+    .cart-list__item--image {
+      position: absolute;
+      left: 20px;
+      top: 50%;
+      transform: translateY(-50%);
+      border: 2px solid #fff;
+      border-radius: 50%;
+      border-width: 4px;
+      width: 60px;
+      height: 60px;
+      .item-Img {
+        width: 60px;
+        height: 60px;
+        border-radius: 42.5px;
+        background-image: url(https://cdn.shopify.com/s/files/1/1148/3974/products/Brandywine-tomatoes-front_compact.jpeg);
+        background-position: center center;
+        background-size: 80px 80px;
+      }
+    }
+    .cart-list__item--name {
+      font-size: 14px;
+      position: absolute;
+      left: 150px;
+      top: 50%;
+      transform: translateY(-50%);
+    }
+    .cart-list__item--quantity {
+      //여기는 아직 아무내용 없음
+      position: absolute;
+      left: 380px;
+      top: 50%;
+      transform: translateY(-50%);
+      .purchase-quantity {
+        display: flex;
+        flex-direction: row;
+        .minus-btn__container,
+        .plus-btn__container {
+          display: flex;
+          align-items: center;
+          justify-content: center;
+          width: 15px;
+          font-size: 20px;
+          color: #ada2a4;
+          padding: 0 5px;
+        }
+        .purchase-input {
+          width: 40px;
+          height: 40px;
+          font-size: 14px;
+          text-align: center;
+          color: #ada2a4;
+          background-color: #fff;
+          border-radius: 3px;
+          outline: none;
+          border: none;
+          font-family: $srf;
+        }
+      }
+    }
+    .cart-list__item--price {
+      position: absolute;
+      left: 495px;
+      top: 50%;
+      transform: translateY(-50%);
+    }
+    .cart-list__item--remove {
+      //버튼 들어갈 자리
+      position: absolute;
+      left: 600px;
+      top: 50%;
+      transform: translateY(-50%);
+    }
+  }
+}
