+++ src/Components/NavForm/LogoContainer.scss
@@ -1,4 +1,4 @@
-@import './var.scss';
+@import '../../Styles/var.scss';
 $s-srf: Raleway, san-serif;
 $srf: Bitter, serif;
