+++ src/Components/NavForm/MainNav.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../Styles/mainNav.scss";
+import "./mainNav.scss";
 import { withRouter } from "react-router-dom";
 
 class MainNav extends React.Component {
@@ -35,14 +35,18 @@ class MainNav extends React.Component {
     return (
       <div className="MainNav">
         <ul className="nav--list">
+          <li className="nav--list__item">
+            <div className="item" onClick={this.goToPage}>
+              Home
+            </div>
+          </li>
           <li
             className="nav--list__item opening"
             onMouseEnter={this.handlePlus}
             onMouseLeave={this.hadleMinus}
           >
-            {/* 여기서 부터 */}
             <div className="item">
-              <span>Home</span>
+              <span>Product</span>
               <input type="ckeckbox" id="plusMinus" />
               <label className="plusMinus">
                 <span className="firstChild"></span>
@@ -58,18 +62,16 @@ class MainNav extends React.Component {
               style={{ display: this.state.listDisplay }}
             >
               <li className="dropdown-list__item" onClick={this.goToPage}>
-                Home ver 1
+                Product
               </li>
-              <li className="dropdown-list__item">Home ver 2</li>
-              <li className="dropdown-list__item">Home ver 3</li>
-              <li className="dropdown-list__item">Home ver 4</li>
+              <li className="dropdown-list__item">Fresh</li>
+              <li className="dropdown-list__item">Frozen</li>
+              <li className="dropdown-list__item">Dried</li>
+              <li className="dropdown-list__item">Liquid</li>
+              <li className="dropdown-list__item">sale</li>
+              <li className="dropdown-list__item">Special</li>
+              <li className="dropdown-list__item">Cooked</li>
             </ul>
-            {/* 숨겨진 애들 끝 */}
-          </li>
-          <li className="nav--list__item">
-            <div className="item" onClick={this.goToPage}>
-              Product
-            </div>
           </li>
           <li className="nav--list__item">
             <div className="item">Blog</div>
