+++ src/Components/NavForm/mainNav.scss
@@ -1,4 +1,4 @@
-@import './var.scss';
+@import '../../Styles/var.scss';
 
 .MainNav {
   min-width: 770px;
@@ -28,6 +28,7 @@
       display: inline;
       margin: 0 5 px;
       position: relative;
+      padding: 0px;
       div.item,
       div.item.opening {
         // width: 100px;
@@ -106,6 +107,7 @@
   .dropdown-list__item {
     text-align: center;
     display: block;
+    height: 35px;
     width: 120px;
     padding: 15px 0 15px 0;
     // margin-top: 15px;
