--- src/Styles/cart.scss
@@ -1,1 +1,0 @@
-@import './var.scss';
